intermediate safety commit: Moved CFGVertex from CFGGenerator to it's own file Split CFGVertex into BytecodeInstruction and AbstractInsnWrapper ... names not set in stone yet Fixed compiler bugs and got branch coverage running (on SimpleTestClass ...) DefUse definetly broken for now
First steps towards a major refactoring of the CFG as a prelude
to the implementation of  efficient computing of the real CDG

Disclaimer:
This is not a working version!
It is even possible this will be rolled back.
If this gets pushed alone then that is not intended!
Contact me if necessary (and ... i guess i'm sorry :D )